Refactor out the redirect url code
